Rename section package to adapter
