+++ app/src/main/java/com/marverenic/music/activity/SearchActivity.java
@@ -22,13 +22,13 @@ import com.marverenic.music.instances.Artist;
 import com.marverenic.music.instances.Genre;
 import com.marverenic.music.instances.Playlist;
 import com.marverenic.music.instances.Song;
-import com.marverenic.music.section.AlbumSection;
-import com.marverenic.music.section.ArtistSection;
-import com.marverenic.music.section.BasicEmptyState;
-import com.marverenic.music.section.GenreSection;
-import com.marverenic.music.section.HeaderSection;
-import com.marverenic.music.section.PlaylistSection;
-import com.marverenic.music.section.SongSection;
+import com.marverenic.music.adapter.AlbumSection;
+import com.marverenic.music.adapter.ArtistSection;
+import com.marverenic.music.adapter.BasicEmptyState;
+import com.marverenic.music.adapter.GenreSection;
+import com.marverenic.music.adapter.HeaderSection;
+import com.marverenic.music.adapter.PlaylistSection;
+import com.marverenic.music.adapter.SongSection;
 import com.marverenic.music.player.PlayerController;
 import com.marverenic.music.view.BackgroundDecoration;
 import com.marverenic.music.view.DividerDecoration;
