+++ app/src/main/java/com/marverenic/music/activity/instance/ArtistActivity.java
@@ -24,13 +24,13 @@ import com.marverenic.music.data.store.ThemeStore;
 import com.marverenic.music.instances.Album;
 import com.marverenic.music.instances.Artist;
 import com.marverenic.music.instances.Song;
-import com.marverenic.music.section.AlbumSection;
-import com.marverenic.music.section.ArtistBioSingleton;
-import com.marverenic.music.section.HeaderSection;
-import com.marverenic.music.section.LibraryEmptyState;
-import com.marverenic.music.section.LoadingSingleton;
-import com.marverenic.music.section.RelatedArtistSection;
-import com.marverenic.music.section.SongSection;
+import com.marverenic.music.adapter.AlbumSection;
+import com.marverenic.music.adapter.ArtistBioSingleton;
+import com.marverenic.music.adapter.HeaderSection;
+import com.marverenic.music.adapter.LibraryEmptyState;
+import com.marverenic.music.adapter.LoadingSingleton;
+import com.marverenic.music.adapter.RelatedArtistSection;
+import com.marverenic.music.adapter.SongSection;
 import com.marverenic.music.lastfm.data.store.LastFmStore;
 import com.marverenic.music.lastfm.model.Image;
 import com.marverenic.music.lastfm.model.LfmArtist;
