+++ app/src/main/java/com/marverenic/music/adapter/EditableSongSection.java
@@ -1,4 +1,4 @@
-package com.marverenic.music.section;
+package com.marverenic.music.adapter;
 
 import android.support.v4.util.ArrayMap;
 import android.view.ViewGroup;
