+++ app/src/main/java/com/marverenic/music/adapter/GenreSection.java
@@ -1,4 +1,4 @@
-package com.marverenic.music.section;
+package com.marverenic.music.adapter;
 
 import android.support.annotation.NonNull;
 import android.support.v4.app.Fragment;
