+++ app/src/main/java/com/marverenic/music/adapter/LibraryEmptyState.java
@@ -1,4 +1,4 @@
-package com.marverenic.music.section;
+package com.marverenic.music.adapter;
 
 import android.app.Activity;
 import android.content.Intent;
