+++ app/src/main/java/com/marverenic/music/fragments/GenreFragment.java
@@ -12,8 +12,8 @@ import com.marverenic.music.JockeyApplication;
 import com.marverenic.music.R;
 import com.marverenic.music.data.store.MusicStore;
 import com.marverenic.music.instances.Genre;
-import com.marverenic.music.section.GenreSection;
-import com.marverenic.music.section.LibraryEmptyState;
+import com.marverenic.music.adapter.GenreSection;
+import com.marverenic.music.adapter.LibraryEmptyState;
 import com.marverenic.music.view.BackgroundDecoration;
 import com.marverenic.music.view.DividerDecoration;
 import com.marverenic.heterogeneousadapter.HeterogeneousAdapter;
