+++ app/src/main/java/com/marverenic/music/fragments/PlaylistFragment.java
@@ -12,9 +12,9 @@ import com.marverenic.music.JockeyApplication;
 import com.marverenic.music.R;
 import com.marverenic.music.data.store.PlaylistStore;
 import com.marverenic.music.instances.Playlist;
-import com.marverenic.music.section.LibraryEmptyState;
-import com.marverenic.music.section.PlaylistSection;
-import com.marverenic.music.section.SpacerSingleton;
+import com.marverenic.music.adapter.LibraryEmptyState;
+import com.marverenic.music.adapter.PlaylistSection;
+import com.marverenic.music.adapter.SpacerSingleton;
 import com.marverenic.music.view.BackgroundDecoration;
 import com.marverenic.music.view.DividerDecoration;
 import com.marverenic.heterogeneousadapter.HeterogeneousAdapter;
